feat(python): Replace usage of ComplexObjectSerializer with DataClassSerializer . * Remove ComplexObjectSerializer from pyfory/__init__.py. * Update _registry.py to use DataClassSerializer directly instead of ComplexObjectSerializer. * Remove ComplexObjectSerializer definition from _struct.py.
